+++ Manager/App.xaml.cs
@@ -336,7 +336,7 @@ namespace Sheepy.Modnix.MainGUI {
       } catch ( Exception ex ) { return Log( ex, "error" ); } }
 
       // Try to detect game path
-      private bool FoundGame ( out string gamePath ) { gamePath = null; try {
+      private bool FoundGame ( out string gamePath ) { try {
          if ( IsGamePath( gamePath = Settings.GamePath ) ) return true;
          foreach ( var path in new string[] { ".", "..", Path.Combine( "..", ".." ) } )
             if ( IsGamePath( gamePath = path ) ) return true;
@@ -349,7 +349,17 @@ namespace Sheepy.Modnix.MainGUI {
       } catch ( IOException ex ) { gamePath = null; return Log( ex, false ); } }
 
       private string SearchRegistry () { try {
-         Log( $"Finding {GAME_EXE} in Windows Game Bar registry" );
+         Log( "Checking Steam registry" );
+         using ( RegistryKey steam = Registry.LocalMachine.OpenSubKey( "SOFTWARE\\WOW6432Node\\Valve\\Steam" ) ) {
+            var path = Path.Combine( steam?.GetValue( "InstallPath" )?.ToString(), "steamapps", "common", "Phoenix Point" );
+            if ( IsGamePath( path ) ) return path;
+         }
+         Log( "Checking Steam App Uninstall registry" );
+         using ( RegistryKey steamPP = Registry.LocalMachine.OpenSubKey( "SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Uninstall\\Steam App 839770" ) ) {
+            var path = steamPP?.GetValue( "InstallLocation" )?.ToString();
+            if ( IsGamePath( path ) ) return path;
+         }
+         Log( "Scanning Windows Game Bar registry" );
          using ( RegistryKey gamebar = Registry.CurrentUser.OpenSubKey( "System\\GameConfigStore\\Children" ) ) {
             foreach ( var game in gamebar.GetSubKeyNames() ) {
                using ( RegistryKey key = gamebar.OpenSubKey( game ) ) {
@@ -368,14 +378,16 @@ namespace Sheepy.Modnix.MainGUI {
          foreach ( var drive in DriveInfo.GetDrives() ) try {
             if ( drive.DriveType != DriveType.Fixed ) continue;
             if ( ! drive.IsReady ) continue;
-            var path = Path.Combine( drive.Name, "Program Files", "Epic Games", "PhoenixPoint" );
+            var path = Path.Combine( drive.Name, "Program Files (x86)", "Steam", "steamapps", "common", "Phoenix Point" );
+            if ( IsGamePath( path ) ) return path;
+            path = Path.Combine( drive.Name, "Program Files", "Epic Games", "PhoenixPoint" );
             if ( IsGamePath( path ) ) return path;
          } catch ( SystemException ) { }
          return null;
       } catch ( Exception ex ) { return Log< string >( ex, null ); } }
 
       internal bool IsGamePath ( string path ) { try {
-         if ( string.IsNullOrWhiteSpace( path ) ) return false;
+         if ( string.IsNullOrWhiteSpace( path ) || ! Directory.Exists( path ) ) return false;
          string exe = Path.Combine( path, GAME_EXE ), dll = Path.Combine( path, DLL_PATH, GAME_DLL );
          Log( $"Detecting game at {path}" );
          return File.Exists( exe ) && File.Exists( dll );
@@ -386,8 +398,12 @@ namespace Sheepy.Modnix.MainGUI {
          CurrentGame.DeleteRootFile( GAME_LOG );
          if ( type == "online" ) {
             if ( CurrentGame.GameType == "epic" ) {
-               Log( "Launching through epic game launcher" );
+               Log( "Launching through Epic Games" );
                Process.Start( Settings.EgsCommand ?? "com.epicgames.launcher://apps/Iris?action=launch", Settings.EgsParameter );
+               return;
+            } else {
+               Log( "Launching through Steam" );
+               Process.Start( Settings.SteamCommand ?? "steam://rungameid/839770" );
                return;
             }
          } else {