+++ Manager/Resources/Changelog.md
@@ -8,7 +8,7 @@ Roadmap: https://github.com/Sheep-y/Modnix/issues/24
 * New: Actions now support "Include".
 * New: Re-implement Action.OnError to properly support "continue", "skip", and "stop".
 * Gui: Mod load order updated to match Modnix 3 phases.
-* Mod: Dlls and action includes must be in same or subfolder relative to the mod.
+* Mod: Action includes must be in same folder as the mod or a subfolder.
 * Log: OS type and version are now logged.
 
 # Version 2.5.3, 2020-07-09